test-requirements.txt
pytest>=3.2.3
pytest-cov>=2.5.1
pytest-mock>=1.6.3
pytest-benchmark>=3.1.1
pytest-flask>=0.10.0
tox>=2.9.1
tox-pyenv>=1.1.0
pylint>=1.8.2
